Changes... :
* fixed cstring::format
* integer types are comtapible with stdint types
* CML_ASSERT macro enables assert
* general I2C.hpp file has I2C_base definition
* I2C and USART properly implemented for L0/L4
* added buffered input option for console
* minor Logger optimizations
* changes in build system

+++ lib/common/integer.hpp
@@ -10,24 +10,24 @@
 namespace cml {
 namespace common {
 
-static_assert(1 == sizeof(char));
-static_assert(2 == sizeof(short int));
-static_assert(4 == sizeof(int));
-static_assert(8 == sizeof(long long int));
+static_assert(1 == sizeof(signed char));
+static_assert(2 == sizeof(signed short int));
+static_assert(4 == sizeof(signed long int));
+static_assert(8 == sizeof(signed long long int));
 
 static_assert(1 == sizeof(unsigned char));
 static_assert(2 == sizeof(unsigned short int));
-static_assert(4 == sizeof(unsigned int));
+static_assert(4 == sizeof(unsigned long int));
 static_assert(8 == sizeof(unsigned long long int));
 
-using int8  = char;
-using int16 = short int;
-using int32 = int;
-using int64 = long long int;
+using int8  = signed char;
+using int16 = signed short int;
+using int32 = signed long int;
+using int64 = signed long long int;
 
 using uint8  = unsigned char;
 using uint16 = unsigned short int;
-using uint32 = unsigned int;
+using uint32 = unsigned long int;
 using uint64 = unsigned long long int;
 
 using byte = uint8;
